Refactor UI networking and improve encapsulation
