+++ bang_py/ui_components/__init__.py
@@ -3,6 +3,7 @@
 from .start_menu import StartMenu
 from .host_join_dialog import HostJoinDialog
 from .game_view import GameView, GameBoard, CardButton
+from .network_threads import ServerThread, ClientThread
 
 __all__ = [
     "StartMenu",
@@ -10,5 +11,6 @@ __all__ = [
     "GameView",
     "GameBoard",
     "CardButton",
+    "ServerThread",
+    "ClientThread",
 ]
-
