+++ bang_py/ui_components/host_join_dialog.py
@@ -33,4 +33,3 @@ class HostJoinDialog(QtWidgets.QDialog):
         buttons.accepted.connect(self.accept)
         buttons.rejected.connect(self.reject)
         layout.addRow(buttons)
-
